requirements.txt
llama-cpp-python==0.2.38
ctransformers==0.2.27
langchain==0.1.5
numpy==1.26.3

